pyproject.toml
[tool.poetry]
name = "aula 3 - análise de dados e gráficos no python"
version = "0.1.0"
description = ""
authors = ["Felipe Santos <fcsantos.ap@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
plotly-express = "^0.4.1"
streamlit = "^1.40.2"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
